Botの回答をEmbedに変更 エラーハンドリングの追加 helpコマンドの実装

+++ bot/__init__.py
@@ -4,6 +4,7 @@ from discord.ext import commands
 import os
 import logging
 from typing import Optional
+from datetime import datetime
 from . import gemini
 from . import myutils
 
@@ -24,6 +25,24 @@ ALLOWED_MIME: set[str] = {
     "audio/ogg",
     "audio/flac", "audio/x-flac",
 }
+MAX_PROMPT_LEN = 1800
+EMBED_SET: dict = {
+    "error": {"title": "Error", "colour": discord.Colour.red()},
+    "answer": {"title": "Answer", "colour": discord.Colour.blue()},
+    "image": {"title": "Image", "colour": discord.Colour.green()},
+    "info": {"title": "Info", "colour": discord.Colour.gold()},
+    "help": {"title": "Help", "colour": discord.Colour.pink()},
+}
+HELP_TEXT = """スラッシュコマンドを通してGEMINI APIを利用できます
+- **/ask**: AIに質問します 画像か音声ファイルを添付できます
+- **/image**: 画像を生成します
+- **/info**: モデルの情報を取得します
+- **/help**: /コマンドの情報を取得します"""
+HELP_EMBED = discord.Embed(
+    title=EMBED_SET["help"]["title"],
+    description=HELP_TEXT,
+    colour=EMBED_SET["help"]["colour"],
+)
 
 logger = logging.getLogger(__name__)
 handler = logging.FileHandler(filename='discord.log', encoding='utf-8', mode='w')
@@ -66,27 +85,50 @@ bot = CUSTOM_AI_BOT()
 @describe(text="質問内容")
 @describe(file="ファイルを添付    画像ファイル 対応形式: PNG/JPEG/WEBP/HEIC/HEIF 音声ファイル 対応形式: WAV/MP3/AIFF/AAC/OGG Vorbis/FLAC")
 async def ask(ctx:discord.Interaction, text:str, file: Optional[discord.Attachment]=None) -> None:
-    await ctx.response.defer(thinking=True)
+    if len(text) > MAX_PROMPT_LEN:
+        await ctx.response.defer(ephemeral=True, thinking=True)
+        await ctx.followup.send(embed=discord.Embed(
+            title=EMBED_SET["error"]["title"],
+            description="質問が長すぎます",
+            colour=EMBED_SET["error"]["colour"]
+            ), ephemeral=True)
+        return
     parts: list[dict] = [{"text": text}]
     file_msg: Optional[discord.WebhookMessage] = None
     if file is not None:
         if file.content_type not in ALLOWED_MIME:
-            await ctx.followup.send("サポートされていないファイル形式です。")
+            await ctx.response.defer(ephemeral=True, thinking=True)
+            await ctx.followup.send(embed=discord.Embed(
+                title=EMBED_SET["error"]["title"],
+                description="サポートされていないファイル形式です",
+                colour=EMBED_SET["error"]["colour"]
+                ), ephemeral=True)
             return
         data = await file.read()
+        await ctx.response.defer(thinking=True)
         file_msg = await ctx.followup.send(file=await file.to_file())
         parts.insert(0, {
             "file_data": {"mime_type": file.content_type, "data": data}
         })
-    response:str = await gemini.generate_text(parts) # type: ignore
+    else:
+        await ctx.response.defer(thinking=True)
+    response: str = await gemini.generate_text(parts) # type: ignore
     header = f"**{ctx.user.display_name}**: {text}\n\n"
-    content = f"**{bot.user.name}**の回答:\n{response}" #type: ignore
-    chunks = await myutils.split_message(header + content)
+    # content = f"**{bot.user.name}**の回答:\n{response}" #type: ignore
+    chunks = await myutils.split_message(response)
     for idx, chunk in enumerate(chunks):
-        if idx == 0 and file_msg is not None:
-            await file_msg.edit(content=chunk)
+        embed = discord.Embed(description=chunk, colour=EMBED_SET["answer"]["colour"]) # type: ignore
+        if idx == 0:
+            embed.set_author(
+                name=bot.user.name, # type: ignore
+                icon_url=bot.user.display_avatar # type: ignore
+            )
+            if file_msg:
+                await file_msg.edit(content=header, embed=embed)
+            else:
+                await ctx.followup.send(content=header, embed=embed)
         else:
-            await ctx.followup.send(content=chunk)
+            await ctx.followup.send(embed=embed)
     
 @bot.tree.command(
     name="image",
@@ -94,21 +136,60 @@ async def ask(ctx:discord.Interaction, text:str, file: Optional[discord.Attachme
 )
 @describe(prompt="生成する画像の説明")
 async def image(ctx: discord.Interaction, prompt: str):
+    if len(prompt) > MAX_PROMPT_LEN:
+        await ctx.response.defer(ephemeral=True, thinking=True)
+        await ctx.followup.send(embed=discord.Embed(
+            title=EMBED_SET["error"]["title"],
+            description="プロンプトがが長すぎます",
+            colour=EMBED_SET["error"]["colour"]
+            ), ephemeral=True)
+        return
     await ctx.response.defer(thinking=True)
     path, text = await gemini.generate_image(prompt)
     if path is not None:
-        await ctx.followup.send(content=f"**{ctx.user.display_name}**: {prompt}", file=discord.File(path))
+        now = datetime.now().strftime("%Y%m%d_%H%M%S")
+        filename = f"{now}.png"
+        file = discord.File(path, filename)
+        embed = discord.Embed(title=prompt, description=text, colour=EMBED_SET["image"]["colour"])
+        embed.set_image(url=f"attachment://{filename}")
+        await ctx.followup.send(embed=embed, file=file)
         os.remove(path)
     else:
-        await ctx.followup.send(content=f"**エラーが発生しました**: {text}")
+        await ctx.followup.send(embed=discord.Embed(
+            title=EMBED_SET["error"]["title"],
+            description="エラーが発生しました",
+            colour=EMBED_SET["error"]["colour"]
+        ))
         
 @bot.tree.command(
     name="info",
     description="Botの情報を確認"
 )
 async def info(ctx: discord.Interaction):
-    text = f"チャットモデル: {gemini.MODEL}\n画像生成モデル: {gemini.IMAGE_MODEL}"
-    await ctx.response.send_message(text)
+    text = f"**チャットモデル**: {gemini.MODEL}\n**画像生成モデル**: {gemini.IMAGE_MODEL}"
+    embed = discord.Embed(
+        title=EMBED_SET["info"]["title"],
+        description=text,
+        colour=EMBED_SET["info"]["colour"],
+        )
+    embed.set_author(
+        name=bot.user.name, # type: ignore
+        icon_url=bot.user.display_avatar # type: ignore
+    )
+    embed.set_footer(text="developed by ellettie")
+    await ctx.response.send_message(embed=embed)
+    
+@bot.tree.command(
+    name="help",
+    description="/コマンド一覧"
+)
+async def help(ctx: discord.Interaction):
+    HELP_EMBED.set_author(
+        name=bot.user.name, # type: ignore
+        icon_url=bot.user.display_avatar # type: ignore
+    )
+    HELP_EMBED.set_footer(text="developed by ellettie")
+    await ctx.response.send_message(embed=HELP_EMBED)
     
 @bot.event
 async def on_ready() -> None:
